Add GitLab pipeline file
The master branch is now protected in GitLab and will only allow Merge
Requests (and reject directly pushing to it).

Whenever the master branch is updated, the GitLab CI will try to
build and publish a new package to the npmjs registry. This will still
need to satisfy the rules for publish-please and will fail if the git
tag is not correct, or a package with that version number already
exists.

After publishing, the CI will for the project will then build and
publish new TypeDocs for the module to the GitLab Pages site at:
https://jych.gitlab.io/gif-provider.